deps/ofi/fabtests/pytest/requirements.txt
pytest
pytest-xdist
pytest-html
pyyaml
retrying
junitparser
